feat: 누적 행동 통계 컴포넌트 분리
